+++ src/app/test-chaemin/page.tsx
@@ -3,10 +3,10 @@ import Button from "@component/components/common-components/button";
 import Modal from "@component/components/common-components/modal";
 import Tag from "@component/components/common-components/tag";
 import CompleteModal from "@component/components/modal/CompleteModal";
-import InterestModal from "@component/components/modal/InterestModal";
-import LoginModal from "@component/components/modal/LoginModal";
-import NicknameModal from "@component/components/modal/NicknameModal";
-import UserInfoModal from "@component/components/modal/UserInfoModal";
+import InterestModal from "@component/components/sign-up/InterestModal";
+import LoginModal from "@component/components/sign-up/LoginModal";
+import NicknameModal from "@component/components/sign-up/NicknameModal";
+import UserInfoModal from "@component/components/sign-up/UserInfoModal";
 import { useState } from "react";
 
 export default function Laboratory() {
